+++ sei-ibc-go/modules/apps/transfer/keeper/relay_test.go
@@ -192,8 +192,7 @@ func (suite *KeeperTestSuite) TestOnRecvPacket() {
 				// relay send packet
 				fungibleTokenPacket := types.NewFungibleTokenPacketData(coinFromBToA.Denom, coinFromBToA.Amount.String(), suite.chainB.SenderAccount.GetAddress().String(), suite.chainA.SenderAccount.GetAddress().String())
 				packet := channeltypes.NewPacket(fungibleTokenPacket.GetBytes(), 1, path.EndpointB.ChannelConfig.PortID, path.EndpointB.ChannelID, path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID, clienttypes.NewHeight(0, 110), 0)
-				ack := channeltypes.NewResultAcknowledgement([]byte{byte(1)})
-				err = path.RelayPacket(packet, ack.Acknowledgement())
+				err = path.RelayPacket(packet)
 				suite.Require().NoError(err) // relay committed
 
 				seq++